Remove old Elasticsearch mapping files - Removed versions 1.0.0, 1.0.1, and 2.0.0 - Renamed mapping_201.json to mapping. From now on we will have just one mapping file - Renamed mapping_201.json in test_data to mapping.json. This file might be removed as well at some point
