--- indexes/repo/mapping_100.json
@@ -1,223 +1,0 @@
-{
-  "repo": {
-    "properties": {
-      "repoID": {
-        "type": "string",
-        "index": "not_analyzed"
-      },
-      "agency": {
-        "type": "nested",
-        "include_in_root": true,
-        "properties": {
-          "name": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "acronym": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "website": {
-            "type": "string",
-            "index": "not_analyzed"
-          },
-          "codeUrl": {
-            "type": "string",
-            "index": "not_analyzed"
-          }
-        }
-      },
-      "status": {
-        "type": "string",
-        "analyzer": "keyword_ci"
-      },
-      "vcs": {
-        "type": "string",
-        "analyzer": "keyword_ci"
-      },
-      "repository": {
-        "type": "string",
-        "analyzer": "keyword_ci"
-      },
-      "name": {
-        "type": "string",
-        "analyzer": "keyword_ci"
-      },
-      "homepage": {
-        "type": "string",
-        "analyzer": "keyword_ci"
-      },
-      "downloadURL": {
-        "type": "string",
-        "index": "not_analyzed"
-      },
-      "description": {
-        "type": "string",
-        "analyzer": "englishfulltext",
-        "fields": {
-          "_fulltext": {
-            "type": "string",
-            "analyzer": "englishfulltext"
-          }
-        }
-      },
-      "events": {
-        "type": "string",
-        "analyzer": "keyword_ci",
-        "fields": {
-          "_fulltext": {
-            "type": "string",
-            "analyzer": "englishfulltext"
-          }
-        }
-      },
-      "tags": {
-        "type": "string",
-        "analyzer": "keyword_ci",
-        "fields": {
-          "_fulltext": {
-            "type": "string",
-            "analyzer": "englishfulltext"
-          }
-        }
-      },
-      "languages": {
-        "type": "string",
-        "analyzer": "keyword_ci",
-        "fields": {
-          "_fulltext": {
-            "type": "string",
-            "analyzer": "englishfulltext"
-          }
-        }
-      },
-      "updated": {
-        "type": "nested",
-        "include_in_root": true,
-        "properties": {
-          "metadataLastUpdated": {
-            "type": "date",
-            "format": "strict_date_optional_time||epoch_millis"
-          },
-          "lastCommit": {
-            "type": "date",
-            "format": "strict_date_optional_time||epoch_millis"
-          },
-          "sourceCodeLastModified": {
-            "type": "date",
-            "format": "strict_date_optional_time||epoch_millis"
-          }
-        }
-      },
-      "contact": {
-        "type": "nested",
-        "include_in_root": true,
-        "properties": {
-          "name": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "email": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "twitter": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "phone": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          }
-        }
-      },
-      "partners": {
-        "type": "nested",
-        "include_in_root": true,
-        "properties": {
-          "name": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          },
-          "email": {
-            "type": "string",
-            "analyzer": "keyword_ci",
-            "fields": {
-              "_fulltext": {
-                "type": "string",
-                "analyzer": "englishfulltext"
-              }
-            }
-          }
-        }
-      },
-      "license": {
-        "type": "string",
-        "analyzer": "keyword_ci",
-        "fields": {
-          "_fulltext": {
-            "type": "string",
-            "analyzer": "englishfulltext"
-          }
-        }
-      },
-      "openSourceProject": {
-        "type": "byte"
-      },
-      "governmentWideReuseProject": {
-        "type": "byte"
-      },
-      "closedProject": {
-        "type": "byte"
-      },
-      "exemption": {
-        "type": "byte"
-      },
-      "exemptionText": {
-        "type": "string"
-      }
-    }
-  }
-}
