+++ services/indexer/term/index.js
@@ -6,7 +6,7 @@ const getConfig = require('../../../config');
 
 // NOTE: dependent on elasticsearch repos being indexed
 
-const ES_REPO_MAPPING = require("../../../indexes/repo/mapping_201.json");
+const ES_REPO_MAPPING = require("../../../indexes/repo/mapping.json");
 const ES_REPO_SETTINGS = require("../../../indexes/repo/settings.json");
 const ES_REPO_PARAMS = {
   "esAlias": "repos",
